requirements.txt
shiny==0.4.0
python-dotenv==1.0.0
openai==1.34.0
matplotlib==3.8.1
pyalex==0.12
tiktoken==0.5.1
pandas==2.1.3
plotly==5.18.0
scipy==1.11.3
scikit-learn==1.3.2
jinja2==3.1.2
